+++ Changelog.md
@@ -8,6 +8,7 @@
 
 - ### Bug fixes
   - [#535](../../issues/535) - BorderBush on bottom of RibbonTabItem (and Ribbon)
+  - [#542](../../issues/542) - InRibbonGallery not reducing properly
   - [#543](../../issues/543) - Using images that can't be found during design time crashes designer  
   A generic "error" image is rendered during design time and an exception is thrown during runtime.
   - [#551](../../issues/551) - "Auto" size for ribbon group box header to support custom font sizes (thanks @chrfin)